# backend/docs/swagger.yaml
basePath: /
definitions:
  models.CategoryStat:
    properties:
      count:
        type: integer
      name:
        type: string
    type: object
  models.CreateEquipmentRequest:
    properties:
      asset_code:
        maxLength: 50
        minLength: 1
        type: string
      category:
        maxLength: 50
        type: string
      last_maint_date:
        type: string
      location:
        maxLength: 100
        type: string
      maint_interval:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - asset_code
    - maint_interval
    - name
    type: object
...
  models.DeleteEquipmentRequest:
    properties:
      lid:
        type: string
    required:
    - lid
    type: object
...
  models.EquipmentDetail:
    properties:
      asset_code:
        type: string
      category:
        type: string
      last_maint_date:
        type: string
      lid:
        type: string
      location:
        type: string
      maint_interval:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  models.EquipmentPublicResponse:
    properties:
      asset_code:
        type: string
      category:
        type: string
      has_active_report:
        type: boolean
      lid:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  models.EquipmentSummary:
    properties:
      faultRate:
        type: integer
      faulty:
        type: integer
      normal:
        type: integer
      pending:
        type: integer
      repairing:
        type: integer
      total:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  models.MonthlyTrend:
    properties:
      faults:
        type: integer
      maintenance:
        type: integer
      monthKey:
        type: string
      name:
        type: string
    type: object
...
  models.StatsResponse:
    properties:
      category_stats:
        items:
          $ref: '#/definitions/models.CategoryStat'
        type: array
      equipment_summary:
        $ref: '#/definitions/models.EquipmentSummary'
      monthly_trends:
        items:
          $ref: '#/definitions/models.MonthlyTrend'
        type: array
    type: object
...
  models.UpdateEquipmentRequest:
    properties:
      asset_code:
        maxLength: 50
        minLength: 1
        type: string
      category:
        maxLength: 50
        type: string
      lid:
        type: string
      location:
        maxLength: 100
        type: string
      maint_interval:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - lid
    type: object
...
info:
  contact: {}
  description: 這是一個整合 PostgreSQL 與 Gin 的設備管理系統。
  title: 設備管理系統 API
  version: "1.0"
paths:
...
  /api/private/equipment:
    delete:
      consumes:
      - application/json
      description: 刪除一個設備紀錄 (僅限管理員)。若該設備仍有關聯維修紀錄則轉為下架/封存，以保留歷史稽核。
      parameters:
      - description: 刪除設備請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteEquipmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功刪除或封存設備
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 設備不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 刪除或下架設備
      tags:
      - private
    patch:
      consumes:
      - application/json
      description: 修改一個設備紀錄 (僅限管理員)
      parameters:
      - description: 修改設備請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEquipmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修改設備
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 設備不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 資產編號已存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 修改設備
      tags:
      - private
    post:
      consumes:
      - application/json
      description: 建立一個新的設備紀錄 (僅限管理員)
      parameters:
      - description: 新增設備請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateEquipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 成功建立設備
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 資產編號已存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 新增設備
      tags:
      - private
  /api/private/equipments:
    get:
      consumes:
      - application/json
      description: 回傳資料庫中未下架設備的完整資訊，支援分頁與關鍵字搜尋 (僅限管理員與維修人員)
      parameters:
      - description: 每頁筆數 (預設 50，最大 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量 (預設 0)
        in: query
        name: offset
        type: integer
      - description: 關鍵字搜尋 (名稱、資產編號、位置，最大 50 字元)
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: 符合條件的總筆數
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.EquipmentDetail'
            type: array
        "400":
          description: 參數錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 獲取所有設備詳情
      tags:
      - private
...
  /api/private/stats:
    get:
      consumes:
      - application/json
      description: 取得儀表板與數據分析頁面所需的設備總數、狀態佔比、分類維修數與近六個月趨勢 (管理員與維修人員皆可存取)
      produces:
      - application/json
      responses:
        "200":
          description: 統計數據
          schema:
            $ref: '#/definitions/models.StatsResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取得系統統計聚合數據
      tags:
      - private
...
  /api/public/equipment:
    get:
      consumes:
      - application/json
      description: 透過 asset_code 取得特定設備的資訊
      parameters:
      - description: 設備資產編號
        in: query
        name: asset_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功回傳設備資訊
          schema:
            $ref: '#/definitions/models.EquipmentPublicResponse'
        "400":
          description: 資產編號必填
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 查詢不到設備
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 查詢設備詳情
      tags:
      - public
...
securityDefinitions:
  BearerAuth:
    description: 請輸入 "Bearer <Your_JWT_Token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"